// web/__generated__/api/models/ProductBody.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Sample Product API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ProductBody
 */
export interface ProductBody {
    /**
     * 
     * @type {string}
     * @memberof ProductBody
     */
    title: string;
    /**
     * 
     * @type {string}
     * @memberof ProductBody
     */
    description?: string;
    /**
     * 
     * @type {number}
     * @memberof ProductBody
     */
    price: number;
    /**
     * 
     * @type {number}
     * @memberof ProductBody
     */
    discountPercentage?: number;
    /**
     * 
     * @type {number}
     * @memberof ProductBody
     */
    stock: number;
    /**
     * 
     * @type {string}
     * @memberof ProductBody
     */
    brand: string;
    /**
     * 
     * @type {string}
     * @memberof ProductBody
     */
    category: string;
    /**
     * 
     * @type {string}
     * @memberof ProductBody
     */
    thumbnail?: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof ProductBody
     */
    images?: Array<string>;
}

/**
 * Check if a given object implements the ProductBody interface.
 */
export function instanceOfProductBody(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "title" in value;
    isInstance = isInstance && "price" in value;
    isInstance = isInstance && "stock" in value;
    isInstance = isInstance && "brand" in value;
    isInstance = isInstance && "category" in value;

    return isInstance;
}

export function ProductBodyFromJSON(json: any): ProductBody {
    return ProductBodyFromJSONTyped(json, false);
}

export function ProductBodyFromJSONTyped(json: any, ignoreDiscriminator: boolean): ProductBody {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'title': json['title'],
        'description': !exists(json, 'description') ? undefined : json['description'],
        'price': json['price'],
        'discountPercentage': !exists(json, 'discountPercentage') ? undefined : json['discountPercentage'],
        'stock': json['stock'],
        'brand': json['brand'],
        'category': json['category'],
        'thumbnail': !exists(json, 'thumbnail') ? undefined : json['thumbnail'],
        'images': !exists(json, 'images') ? undefined : json['images'],
    };
}

export function ProductBodyToJSON(value?: ProductBody | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'title': value.title,
        'description': value.description,
        'price': value.price,
        'discountPercentage': value.discountPercentage,
        'stock': value.stock,
        'brand': value.brand,
        'category': value.category,
        'thumbnail': value.thumbnail,
        'images': value.images,
    };
}